v3.0.0.5
Changes for this release:

When the user enters a crypto key, we now ask that they re-enter it a second time for confirmation.  In addition, in the final output displayed to the screen, we display the original key in addition to the scrambled or unscranbled key.

Much as we did for the crypto key, the user must now re-enter the BitLocker recovery key or the scrambled key as confirmation to make sure that it was not mistyped. The crypto key is no longer limited to 12 digits maximum. It can now be up to 48 numerical digits in length

We have added the ability to output information in 3 ways:

1) Output ALL information including the original recovery key, the scrambled key, and the crypto key to a file.
2) Output ONLY the scrambled key to a file.
3) Create both files with all info and a file that contains only the scrambled info.

Note that in all three cases above, if the user enters a comment, that will also be saved to the file(s).

Added the ability to save the BitLocker Drive Identifier when saving output to files. This will make confirmation of the correct drive easier.